edit profile badges
